Strip out DataSource and related complications
To handle out-of-place saved data, let's just record provenance in the
metadata in some standard format and transforms in the DataSet. And if
we want to support on-demand loading, we can do that with individual
recspans instead of grouping together all data from the same file. So
DataSource isn't really needed and it makes things more complicated.