Turn the old check_transforms utility into a one-off test, and fix it.
We no longer allow for random extension classes to handle implementing
the dataset .transform method, so no need for a utility for testing
them. But the test still is useful -- or at least, it is after being
fixed so that it actually makes a copy of the dataset when it tries
to!